Move seeds from own package to tasks
