+++ nuxt.config.ts
@@ -1,4 +1,7 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
+import fs from "node:fs/promises"
+import path from "node:path"
+
 export default defineNuxtConfig({
   modules: [
     "@nuxt/eslint",
@@ -54,6 +57,25 @@ export default defineNuxtConfig({
     plugins: {
       tailwindcss: {},
       autoprefixer: {},
+    },
+  },
+
+  hooks: {
+    "nitro:build:public-assets": async (nitro) => {
+      // Copy Prisma schema and migrations to the output directory
+      const prismaInDir = path.join(__dirname, "prisma")
+      const prismaOutDir = path.join(nitro.options.output.dir, "prisma")
+      await fs.cp(prismaInDir, prismaOutDir, { recursive: true })
+
+      // Copy Prisma schema engine to the output directory
+      const prismaEnginesInDir = path.join(__dirname, "node_modules", "@prisma", "engines")
+      const prismaEnginesOutDir = path.join(nitro.options.output.serverDir, "node_modules", "@prisma", "engines")
+      const prismaEnginesFiles = await fs.readdir(prismaEnginesInDir)
+      for (const file of prismaEnginesFiles) {
+        if (/^schema-engine-.+$/.test(file)) {
+          await fs.cp(path.join(prismaEnginesInDir, file), path.join(prismaEnginesOutDir, file))
+        }
+      }
     },
   },
