+++ package.json
@@ -16,9 +16,7 @@
     "test": "npm run test:unit && npm run test:e2e",
     "test:unit": "prisma generate && vitest run ./tests/unit",
     "test:unit:coverage": "prisma generate && vitest run --coverage ./tests/unit",
-    "test:e2e": "playwright test",
-    "migrate": "dotenv -e .env.development prisma migrate dev",
-    "migrate:seed": "dotenv -e .env.development prisma db seed"
+    "test:e2e": "playwright test"
   },
   "dependencies": {
     "@formkit/drag-and-drop": "^0.3.3",
@@ -60,6 +58,8 @@
     "@nuxtjs/tailwindcss": "^6.13.1",
     "@playwright/test": "^1.50.0",
     "@prisma/client": "^6.3.0",
+    "@prisma/internals": "^6.3.0",
+    "@prisma/migrate": "^6.3.0",
     "@tailwindcss/typography": "^0.5.16",
     "@types/bcrypt": "^5.0.2",
     "@types/node": "^22.12.0",
@@ -68,8 +68,8 @@
     "@vue/test-utils": "^2.4.6",
     "autoprefixer": "^10.4.20",
     "dotenv-cli": "^8.0.0",
-    "nodemailer": "^6.10.0",
     "happy-dom": "^16.7.3",
+    "nodemailer": "^6.10.0",
     "nuxt-file-storage": "^0.2.9",
     "nuxt-nodemailer": "^1.1.2",
     "playwright-core": "^1.49.1",
