+++ server/plugins/migrate-and-seed.ts
@@ -1,0 +1,4 @@
+export default defineNitroPlugin(async () => {
+  await runTask("db:migrate")
+  await runTask("db:seed")
+})
